Widen encoded minimizers to longs instead of ints, enabling up to 32 bp
Make MinimizerPriorities polymorphic, while still making it possible to broadcast MinSplitter, by doing the following:
  * Remove MinimizerPriorities type parameter (always Long)
  * Parameterize MinSplitter on MinimizerPriorities
  * Encoder of MinSplitter is obtained using a global lookup method

+++ src/main/scala/com/jnpersson/discount/hash/PosRankWindow.scala
@@ -29,7 +29,7 @@ package com.jnpersson.discount.hash
  * @param motifRanks Array of motif priorities at the positions in the underlying read where the full motif can first
  *                   be read (e.g. position 4 for a 5-length motif occupying positions 0-4).
  */
-final class PosRankWindow(m: Int, k: Int, val motifRanks: Array[Int]) {
+final class PosRankWindow(m: Int, k: Int, val motifRanks: Array[Long]) {
   import MinSplitter.INVALID
